Update census generation
It now takes zones, additional information in Members and tagged ports
in the service rumors into account.

Basically if the service is accessible to us (it either is in the same
zone as us or it is exposed in our zone) then we put the information
about the service in the census. Otherwise it is ignored or dropped.

Signed-off-by: Krzesimir Nowak <krzesimir@kinvolk.io>