Add a simple file watcher
If all the file watcher needs is to watch a single file and have a
check function if there were some filesystem events on the file then
it could use the simple file watcher.

The code is based on the peer watcher and the watcher will be using it
in the followup commit.

Signed-off-by: Krzesimir Nowak <krzesimir@kinvolk.io>