Add additional addresses file watcher
The watcher expects the watched file to be a toml file with contents like:

```
[sup.tag1]
address = "1.2.3.4"
swim_port = 1234
gossip_port = 4321

[sup.tag2]
swim_port = 2345
gossip_port = 5432

[svc.nginx.tag1]
port = 3456
ssl-port = 3457

[svc.redis.tag1]
port = 4567

[svc.redis.tag2]
port = 7654
```

This file is a source of additional information of a Member and a
source of tagged port sets for the service rumor. The file has two
kinds of headers: "sup.<TAG_NAME>" and "svc.<SVC_NAME>.<TAG_NAME>".

The "sup" header needs to contain at least "swim_port" and
"gossip_port" fields, the "address" field is optional, though I don't
know if it would be better to make it required - I'm not sold on the
idea of address discovery…

The "svc" header contains names of the ports (taked from default.toml,
for example) and their corresponding values.

From the example above, we can say that nginx is available under
1.2.3.4:3456 for unencrypted connections and under 1.2.3.4:3457 for
encrypted connections.

Signed-off-by: Krzesimir Nowak <krzesimir@kinvolk.io>