Extend service rumor with additional ports information
Services to be usable outside the NAT need to be exposed. We can
assume that the address of the exposed service is the same as the
address of the supervisor, but we can not assume the same for the
ports that the services expose. So the service rumor needs to contain
some information about sets of exposed ports. Each set has a tag which
needs to match a tag in the additional information in the Member
struct. The set consists of name-port pairs, where name of a port
needs to match the name used in the service configuration (user.toml
or default.toml). The supervisor will be able to use that information
to generate a census with proper addresses and ports for services that
are behind a NAT and are exposed.

Signed-off-by: Krzesimir Nowak <krzesimir@kinvolk.io>