Use proper Message implementation for Membership rumor
Not sure why, but `Membership` was implementing its own `from_bytes`
and `write_to_bytes` methods, rather than implementing the `Message`
trait, and thus inheriting the implementations of those functions that
all the other rumors use.

The custom implementations weren't actually compatible with each
other, in that you couldn't roundtrip a Membership rumor through
them. This meant that when restarting a Supervisor, we wouldn't read
in the `Membership` rumors from our binary `rst` file "dump" of
Butterfly's state. This causes all sorts of issues, because all the
other rumors *would* roundtrip, giving an inconsistent view of the
world.

A test was added to ensure the rumor can round-trip itself correctly.

Signed-off-by: Christopher Maier <cmaier@chef.io>