store control gateway address in file
Resolves https://github.com/habitat-sh/habitat/issues/5390

This change creates a new `CTL_GATEWAY_ADDRESS` file under `sup_root`.
The purpose is to improve the UX when the Supervisor was started with
`--listen-ctl <LISTEN_CTL>`. Instead of having to specify `--remote-sup
<REMOTE_SUP>` on every `hab sup *` command, hab will now first attempt
to read the ctl gateway address from file on disk.

This scenario is especially common in Docker exported hab pkg
situations where users might have multiple hab services running in
docker containers, on one Docker Host. In order to do so, they must
customize the `--listen-ctl` address of each Supervisor. In each
container, users must follow up with `--remote-sup` when talking to the
local supervisor, which is cumbersome and error prone (with users often
not specifying the correct address and talking to a different
supervisor!).

Signed-off-by: Jeremy J. Miller <jm@chef.io>