Fix pingreq in multizone setup
We send ping messages only to supervisors that are directly reachable
by us, that is - to supervisors that are in the same zone as us or are
exposed in our zone. So, same should happen for pingreqs. But
currently the code was picking just any member, without any
consideration about reachability.

Now, when need arrives, we try to gather pingreq targets that are in
our zone too. That's a bit stricter than direct reachability, because
it excludes supervisors that are exposed to our zone.

The reason for using stricter condition is to avoid locking both zone
list and member list at the same time, since that may lead to lock
ordering issues. To be able to check if two members are directly
reachable we need to lock zone list. Calling with_pingreq_targets
locks its members lock. To overcome it, I could, for example, pass a
zone list lock to the with_pingreq_targets function so it would lock
the zone list after doing a copy of members (which requires
locking). But I found that solution after writing the current one and
I like the current one more.

Signed-off-by: Krzesimir Nowak <krzesimir@kinvolk.io>