Persist incarnation numbers
A core component of SWIM is the incarnation number, which allows
members to determine which membership status updates are "more
recent". In the absence of persisting these numbers, however, a member
that is rejoining a network will start with an incarnation number of
0, which may be dramatically behind what the rest of the network may
consider to be "current". As a result, the member may not be
recognized as having returned for a potentially long time as it tries
to refute every other member of the network.

Here, we introduce the `IncarnationStore`, an interface over
persisting and retrieving an incarnation number from
disk (`/hab/sup/default/data/INCARNATION`). Each time a Butterfly
server must increment its incarnation number, we write the new value
out to disk. When a server restarts, it reads the value stored on disk
and uses that (rather than 0) to start (it actually increments it by
1, since it is now the next incarnation of itself). If the file is not
present (as will happen with any previous Supervisors without this
feature), it will be created and seeded with an initial value of 0 (in
keeping with the current behavior).

The file itself is a simple text file containing the current
incarnation number.

On startup, if the file is present, but cannot be read or contains
invalid data, the Supervisor will refuse to start. Once started, any
failures to persist incremented incarnation numbers will not bring the
Supervisor down, but will log warnings, allowing operators to
intervene.

Signed-off-by: Christopher Maier <cmaier@chef.io>