Try to reduce a number of running threads in nat emulation
Butterfly server and the NAT emulation already spawn way too many
threads, so let's try limiting them a bit. The NAT emulation spawns
two additional threads for every butterfly server. The threads do the
message routing, so this is something unnecessary after the test has
finished running.

This may result in a bunch of messages about being unable to send a
message to be printed, though.

Signed-off-by: Krzesimir Nowak <krzesimir@kinvolk.io>