Drop uuid from variable names
The differentiation between, say, "zone_id" and "zone_uuid" came from
before the rebase of the code on top of prost. "zone_id" used to be a
"String" and "zone_uuid" used to be "BfUuid". It was like that,
because before using prost, we were using generated data structures
and those used "String" for the zone ID. But we sometimes needed to
use "BfUuid", so to have that we used to have a code like `let
zone_uuid = BfUuid::must_parse(&zone_id)`. After rebasing to prost, we
are using a convenient data structures which use "BfUuid" for zone ID,
so all "zone_id" variables became "BfUuid" too and no "must-parsing"
happen any more. But then we had an inconsistent naming of variables.
Finish the transition by getting rid of "uu" in the "uuid" in variable
names.

Signed-off-by: Krzesimir Nowak <krzesimir@kinvolk.io>