Drop unused variable
We had a pair consisting of a zone and its ID as BfUuid. This is not
necessary anymore as we can get the ID as "BfUuid" from zone
directly. Previously we could get the ID from zone only as a "String",
which then would need to be parsed to become a "BfUuid", which added
yet another panic or failure point in the code.

Signed-off-by: Krzesimir Nowak <krzesimir@kinvolk.io>