Send an ack from ack handler a bit later
We usually try to send information about our own zone when we send
pings or acks. But there is one scenario where we haven't been doing
it. The case is when we received an ack from some server and as a
result of processing it we assume sender's zone as our own. We want to
notify the sender about the fact. If we send the ack message to the
sender too early then our message won't contain information about our
new zone, because we haven't yet added all the zone information to our
zone list. And that's needed, because we populate the ack message with
the zone information from the zone list.

This should silence a warning about getting an ack message without
information about the sender's zone.

Signed-off-by: Krzesimir Nowak <krzesimir@kinvolk.io>