Fix an infrequent panic in zone handling
The code that was panicking was trying to split a set of zones into
separate sets of aliases. There was a corner case where I didn't do a
proper clean up, that lead to trying to get a maximum element from an
empty vector.

The commits does the cleanup and, as an additional measure, stops
using `unwrap()` on an `Option` value to avoid panics.

Signed-off-by: Krzesimir Nowak <krzesimir@kinvolk.io>