Force sending pings from the initial members list
Let's assume that have servers A, B and C, where A has B in its
initial members list and B has C in its initial members list. This
commit tries to fix the case where A sends a ping to B and B puts A in
its members list before B even does its first check on the initial
members list. If this happens, then B will find out that it already
has minimal required number of members in its members list, so it
won't sent any pings to the members in its initial members list. This
results in C never being pinged.

This is a very unlikely scenario in typical cases where supervisors
are running on separate hosts. But the same scenario is quite likely
during the tests, when we run multiple servers that talk to each other
using channels instead of sockets - the number of threads involved
there may cause a scenario that the outbound thread gets its share of
computing time quite late, after an inbound thread processed the ping
and filled the members list.

Signed-off-by: Krzesimir Nowak <krzesimir@kinvolk.io>