Get rid of address discovery
Drop the address discovery for now. I have a feeling that it will
interfere with figuring out the direct communication with sibling
zones and with extending the additional address information with
hostname or http or ctl gateway addresses.

The discovery was quite lame anyway - the only thing it could discover
was an address and the discovery was made based on port numbers
used. Basically the server needed to know exposed port numbers from
the beginning and used them to match the address from the SWIM message
with an additional address that has an unset address.

If we want to readd the discovery then more thought needs to be put
into it. Also, it exists in the git history, so the code is not lost.

Signed-off-by: Krzesimir Nowak <krzesimir@kinvolk.io>