Stop manually manipulating the PATH in bintray-publish package
This package was initially built long ago, before `hab pkg exec` would
manage your `PATH`  contents based on `pkg_deps` on your  behalf. As a
result, this package did that by hand.

The recent removal of studio image uploading from this package (#5560)
removed a portion of this `PATH` management, but inadvertently left
part behind. As a result, the script was overwriting the correct,
`hab`-managed `PATH` with an incomplete one from package build
time. This manifested as an inability to find the `basename` command.

This commit completes the removal of the custom `PATH` management
code, letting `hab pkg exec` do all the heavy lifting for us.

Signed-off-by: Christopher Maier <cmaier@chef.io>