Refactor zone handling
The refactor was to ensure that ZoneList is really at least locally
consistent. The local consistency is about having all the aliases of
our current zone, all the aliases of all the children of our current
zone and all the aliases of the parent of our current zone consistent.

This means that all the aliases of our current zone have the latest
alias (a successor) of our current zone's parent. Conversely, this
makes sure that all aliases of our current zone's parent have latest
alias of our current zone as one of their children.

Similar goes with the children of our current zone.

The "locally" part of the "locally consistent" attribute is about
caring only about consistency of all aliases of our current zone, of
all aliases of all of our current zone's children and of all aliases
of our current zone's parent. For further descendants or ancestors, we
care about having only their aliases consistent.

The reason behind such a selectiveness is that pings only go between
supervisors in the same zone or neighbouring zones.

Signed-off-by: Krzesimir Nowak <krzesimir@kinvolk.io>