Only log about satisfied binds if they were previously unsatisfied
Since we validate binds each time the census changes, and since the
expected normal state of a service is to have its binds satisfied,
saying that a bind is satisfied each time is pointless spamming of the
logs.

Instead, we'll only log about bind satisfaction whenever they were
previously unsatisfied, which is more interesting anyway.

Signed-off-by: Christopher Maier <cmaier@chef.io>