Purge "rumor heat" information for departed members
This should allow us to do a degree of cleanup of old rumor
information. Absent this, large and long-running Habitat networks can
accumulate significant amounts of wasted memory on what is effectively
useless historical information.

The `RumorHeat` data structure effectively maps a rumor to a mapping
of members to the number of times the member has seen the rumor. This
allows us to know who's seen a rumor already, and who hasn't.

Now, when a member departs the ring, we'll remove any of the
"member-to-number-of-times-they've-seen-the-rumor" mappings for that
member.

Additionally, we'll remove _all_ such mappings for any rumors about
_services_ hosted on that member. After all, the member is going away,
so all of its services will too.

We aren't removing the membership rumors, though, because that's how
we pass the word that the member is departed.

Future work may add more cleanups and rearrangements.

Signed-off-by: Christopher Maier <cmaier@chef.io>