Store port values as u16 in Member struct
This is more consistent with the rest of the local representations. Since
protobuf doesn't support 16-bit ints, we have to convert on that boundary, but
we should be checking values there and avoiding unsafe casting elsewhere.

AtomicUsize instances would be converted to AtomicU16s, but integer_atomics
isn't stable yet, so we use Mutex<u16>.
See https://github.com/rust-lang/rust/issues/32976

Signed-off-by: Jon Bauman <5906042+baumanj@users.noreply.github.com>