Reduce excessive disk I/O when persisting rumors
The Butterfly server writes all rumors to disk periodically for
durability. It also writes rumors to disk in the `Drop` implementation
of the `Server` struct. While this could be used to implement a "save
on shutdown" behavior, it also ended up causing a worst-case behavior
of writing the rumors out five times _per second_. This is because a
copy of the Butterfly server is passed into each of the five
short-lived worker threads that are spawned every second for pushing
rumors out to network members. When these threads exit, they drop
their copy of the Butterfly server struct, triggering a write.

These worker threads are started when there are new rumors to send
out, so the more active the network, the more frequently these writes
would happen. With larger networks, there are both more members to
send rumors to, as well as a larger number of rumors to persist. Thus,
large and active networks would generate scenarios of excessive disk
load.

Here, we simply remove the offending `Drop` implementation, thereby
eliminating the 5x/s worst case scenario. For persistence, we instead
rely on the existing periodic write of once every 30s. Additionally,
we add an explicit call to persist the data at Supervisor shutdown
time.

Finally, the logging message for a successful disk write is now
`info`, rather than `debug`, since this is an important event to
notify about.

Signed-off-by: Christopher Maier <cmaier@chef.io>