Add fs::AtomicWriter and fs::atomic_write
The Habitat code has a number of places where it writes to a temporary
file and then renames the file. The goal of this pattern is to achieve
an atomic write to the file to prevent concurrent readers from seeing
incomplete versions of this file.

However, getting this pattern right can be tricky. For example, the
temporary file must be created in the same directory to ensure an
atomic rename is possible and the temporary file must by fsync'd to
disk before the rename to protect against seeing partial files in some
powerloss scenarios.

The AtomicWriter is an attempt to help us get this pattern right
across the code base.

Signed-off-by: Steven Danna <steve@chef.io>