Simplify elections code by replacing ServiceGroup structs with strings
This should have no effect on behavior, it's a simplification that came out of
auditing the elections code.

The change is fairly seamless because ServiceGroup is String wrapper, but for
the purpose of elections, it's only used as an opaque identifier. Using a
string directly allows for simplification of code where the complexity is
appropriate for validating user input that's unnecessary for data that either
comes from protocol messages or existing ServiceGroup instances.

See https://github.com/habitat-sh/core/blob/master/components/core/src/service.rs#L85

Signed-off-by: Jon Bauman <5906042+baumanj@users.noreply.github.com>