Abstract the logging of terminated hook processes
When `ExitStatus#code()` returns `None`, it doesn't _just_ mean the
process exited without a status code, it means it was terminated by a
signal.

Since this can only happen on Unix platforms, we can use the
`ExitStatusExt` trait to extract _which_ signal the process was
terminated by.

To capture the platform-dependence of this behavior, and to not have
to repeat ourselves everywhere, there's a new
`output_termination_message` function in the `Hook` trait that we can
use throughout the code.

Signed-off-by: Christopher Maier <cmaier@chef.io>