Remove panic by moving problem code out of constructor
The root of the panic is due to us trying to serialize a
`Toml::Value::Table` to a Vec<u8> inside a constructor where we can't
handle the error cleanly

This commit pulls the TOML parsing out of Service::new() into
Cfg::to_exported() so we can properly handle the Result<> on toml parsing

Note: this does not yet handle the underlying issue which is that the
TOML serializer doesn't handle tables well.  In particular it requires
all keys with non-table values must be emitted first.  This is not the
default

See https://github.com/alexcrichton/toml-rs/blob/master/src/ser.rs#L7-L12

Signed-off-by: James Casey <james@chef.io>