Remove check_health_of and check_health_of_by_id
The interface of these functions was prone to subtle errors. Specifically,
returning a bool made it easy to overlook the case where the member's
health was unknown. Checking for equality with health_of_by_id is nearly as ergonomic, and considerably less likely to overlook handling this case
correctly. Additionally, it makes for more understandable test output.

Signed-off-by: Jon Bauman <5906042+baumanj@users.noreply.github.com>