Remove minimum 3 member requirement from quorum check
Resolves https://github.com/habitat-sh/habitat/issues/4759

The purpose of quorum is to prevent multiple elections occurring on
different sides of a network partition leading to split-brain. This doesn't
require a minimum of 3 members to work. Also, there's no risk of a tie in
the election process, since the merge operation breaks ties with the member
ID, which is guaranteed to be unique (or else the consequences would be
much worse than split brain).

Signed-off-by: Jon Bauman <5906042+baumanj@users.noreply.github.com>