[hab] Avoid install failure when running concurrently
Commit b688d8ae25d9f38b4053fd29372e2733c9254327 changed the install
process to use a temporary directory + rename strategy. The goal of
that change was to avoid incomplete package installations. However,
that change had an unintended consequence of breaking concurrent
installs.  While concurrent installs were not previously not
guaranteed to work, they often would by chance; now, they never work
because rename(2) fails when the target is a non-empty directory.

This change checks if the package we are installing happens to exist
after an error from rename(2). If it does, we proceed as if there were
no error. If it doesn't we raise the error.

I've opted not to catch a specific error here to avoid platform
specific code. Even on linux alone the manual page for this case says
we might get ENOTEMPTY or EEXIST. I can be convinced we should
check for a specific error though.

Signed-off-by: Steven Danna <steve@chef.io>