Parameterize the rumor sharing limit
Further reading of the "Epidemic Algorithms for Replicated Database
Maintenance" paper reveals that we may not be managing our "rumor
heat" correctly. The paper describes many variations of "cooling" a
hot rumor, all of which speak in terms of sending out a given rumor to
the entire network as a whole, rather than tracking the number of
times each rumor is sent to every individual in the network.

To begin to investigate the impact of bringing our approach more in
line with the documented algorithm, we are making the number of times
we send a rumor to any individual Supervisor tunable using the
environment variable `HAB_RUMOR_SHARE_LIMIT`. This should be
considered an "undocumented" variable, and is only being exposed for
the purposes of internal testing; general Habitat users shouldn't be
using it.

Right now, testers would only ever want to set this variable to `1`,
meaning that each rumor is sent to every other Supervisor once, rather
than the current default of twice. Doing so ought to cut overall
network traffic in half. (It should be noted that each Supervisor
still stands to receive any given rumor as many times as there are
Supervisors in the network, so there is still a considerable amount of
redundancy.)

Signed-off-by: Christopher Maier <cmaier@chef.io>