Only start services that are runnable
Before writing out new service spec files check that the service is
actually runnable. We do the checks in `manager::command` so we
never actually trigger an attempt to start the service (which gives
unhelpful user messages).

We still need to install the package and it's dependencies in order
to check if it's runnable.  We don't clean up those packages

NOTE:  This code doesn't cover the code paths covering the composite
use cases as they're going to be removed.

Signed-off-by: James Casey <james@chef.io>