Update core and http-client crate to use rust 2018 edition idioms
All these changes were made automatically with
cargo fix --edition-idioms --broken-code
With the exception of retaining the "extern crate crypto as rust_crypto;
declaration to resolve ambiguity. The alternative syntax "use ::crypto" is
currently changed to "use crypto" by rustfmt, so we'll have to use the old
way for now.

The win-users crate will be converted in a separate commit

Signed-off-by: Jon Bauman <5906042+baumanj@users.noreply.github.com>