More simplification of previous_peer thanks to help on rust-lang.slack.com
Using by_ref allows us to avoid the creation of the Vec with collect, but
still retain the alive_members iterator to call `last` in the event we need
to wrap around for the "previous" member. When we do find a match, we use
Option's `or_else` method to avoid iterating through the rest of
alive_members unnecessarily.

Signed-off-by: Jon Bauman <5906042+baumanj@users.noreply.github.com>