Convert all of the launcher crates to use prost.
This is the last of the crates in the habitat repo to be converted from
protobuf to prost.  Note that unlike the supervisor and
butterfly, I decided against including the generated code in the repo.

It's also worth noting that unlike our approach with the protobuf crate,
we're not actually using the structs generated by prost except at the
edges. Prost rightfully uses Options to represent optional fields and
since all of our protobuf messages have all optional fields, the
ergonomics of using Prost generated structs is sub-optimal. For this
reason, we have our own set of internal structs that mirror the Prost
ones that we use in the launcher and launcher-client crates, and
launcher-protocol is the only crate that knows about the Prost generated
structs.

Signed-off-by: Josh Black <raskchanky@gmail.com>