Don't copy the entire member list when choosing PingReq values
When choosing PingReq targets for SWIM, we want to randomly choose 5 members
which are:
1. Alive
2. Not the the member doing the sending

Previously we cloned the entire member list, shuffled it, filtered out the
members that didn't satisfy (1) or (2), and then took the first 5 values.

To avoid unnecessary allocation and copying, we can instead use
rand::seq::IteratorRandom::choose_multiple to allocate space for and copy
only the 5 values we need to return. Additionally, we can return member
references rather than owned values to further reduce overhead.

Signed-off-by: Jon Bauman <5906042+baumanj@users.noreply.github.com>