Generate protobuf code in Supervisor and Launcher package builds
With the transition to always-compiled Protobuf code in our launcher
protocol crate (see #6105), we need to add the Protobuf compiler to
our package build process.

I suspect this works unmodified on Windows because the Habitat build
process is necessarily less of a "clean room" affair. The
`prost_build` crate includes a `protoc` binary that it can use, but
this doesn't work in our Linux builds because it's linked to
non-Habitat libraries.

Signed-off-by: Christopher Maier <cmaier@chef.io>