Remove txn_id from Launcher protocol
Recent refactoring of the Launcher protocol crate (see #6105) has
unearthed the fact that the `txn_id` field in the Launcher protocol
envelope was never meaningfully used. That refactoring ended up
treating an Envelope without a `txn_id` as invalid, which ended up
breaking compatibility because nothing actually sent an envelope with
a `txn_id` in the first place.

Since `txn_id` was never actually used, it is now removed. The field
name and number are now treated as reserved so they can never be
reused in the future.

Signed-off-by: Christopher Maier <cmaier@chef.io>