Use AtomicWriter from habitat core
This makes use of the new AtomicWriter in habitat core. The majority
of these call sites were not fsync'ing their related files before the
rename. This mean that in the case of powerloss it was possible to see
partially written files after reboot even if this function returned
without error.

Signed-off-by: Steven Danna <steve@chef.io>