Consolidate some code
Without `txn_id`, we can remove a bit more code.

First, `NetTxn#build_reply` is unnecessary, since it is functionally the same
as `NetTxn::build`, so it can be removed.

Second, without `NetTxn#build_reply`, the `reply` function is exactly
the same as `send`, so _it_ can be removed as well.

Signed-off-by: Christopher Maier <cmaier@chef.io>