Add Launcher build stages to the release pipeline
The pipeline will now ask if we need to build new Launcher artifacts
for this release. If we do, build steps will automatically be added;
if we don't, we'll just add the currently stable Launcher artifacts to
the release channel.

This adds automation that previously did not exist, and also *removes*
the manual step where we had to supply fully-qualified identifiers for
all Launcher packages we wanted to promote into the release
channel. Wins all around!

Signed-off-by: Christopher Maier <cmaier@chef.io>