Separate running rustfmt from installing it.
If someone discovers that CI fails because their code wasn't formatted
properly, it'd be nice to have an easy way to install the same version
of rust nightly that CI is using. The existing script for doing this is
tightly married to BuildKite, so this separates the installation logic
from the BuildKite logic to allow developers to run
install-nightly-rustfmt.sh on their workstations.

Signed-off-by: Josh Black <raskchanky@gmail.com>