Add ShutdownSignal, ShutdownTimeout types
`ShutdownSignal`, available on Unix platforms, is a specialization of
`Signal` which, among other things, allows us to encode parsing and
default logic for a specific use case of a `Signal`.

Similarly, `ShutdownTimeout` encodes as a type what was previously a
magic number constant: the number of seconds to wait after terminating
a process before forcibly killing it.

These types will form the foundation for exposing these concepts to
users via the commandline.

Signed-off-by: Christopher Maier <cmaier@chef.io>