Use subshells and unset to actually unset environment variables
https://github.com/habitat-sh/habitat/pull/6278 made a change to
environment variable handling around the channel that `hab` uses to
install packages. As part of this, we no longer treated an empty
string as being synonymous with "unset".

This, in turn, revealed places in our shell code where the intention
was to unset environment variables, but in practice, they were set to
empty strings.

This PR changes our `_hab` wrapper function in our various studio
implementations to run in a subshell, using POSIX `unset` to remove
variables from the environment completely. Busybox's `env` also
supports the `-u` syntax, which we could have used, but we've opted to
keep with POSIX whenever possible.

Big thanks to @baumanj for this solution.

Signed-off-by: Christopher Maier <cmaier@chef.io>