Make Linux Kernel 2 validation just a little bit easier
To validate `x86_64-linux-kernel2` target releases, we need a Linux
system with an old enough kernel. We have a `Vagrantfile` that does a
basic setup of a CentOS 6 machine, but there are extra steps you need
to take to run services, build plans, and so on.

This PR adds a bit more automation to help with some of that. There
are still manual steps (namely, getting a release candidate `hab`
binary on the box), but this change makes things a lot easier.

The changes are:

* Add a `hab` user and group

  If you want to run a service to test things out, you're probably
  going to need to have those anyway.

* Check out the core-plans repository

  If you want to build a package, you're going to need some plans!

* Install secret keys

  When you want to build a package, you'll need signing keys. When
  provisioning the VM, we'll grab the secret key that corresponds to
  the values of `HAB_ORIGIN` and `HAB_AUTH_TOKEN` in your
  workstation's environment. When you build in the VM, just set
  `HAB_ORIGIN` in the environment of the `hab pkg build` call.

  To do this, we also install the current stable `hab` binary. When
  validating a release candidate, you'll still need to manually get
  _that_ binary on disk. Having the stable binary around helps make
  this a more self-contained system, and also makes this VM a bit more
  useful outside of pure release candidate validation purposes.

* Added some additional documentation around all this.

  Because who doesn't love documentation?

Signed-off-by: Christopher Maier <cmaier@chef.io>